django_forum/requirements..txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
psycopg2>=2.9,<3.0  
django-cors-headers>=3.7,<4.0  
drf-yasg>=1.20,<2.0  
